Refactor : dto 수정
